+++ src/test/kotlin/dev/turingcomplete/intellijbytecodeplugin/openclassfiles/OpenClassFilesTaskTest.kt
@@ -45,7 +45,7 @@ class OpenClassFilesTaskTest(
   @Test
   fun testOpenFile() {
     val actualClassFilesReadyToOpen = mutableListOf<VirtualFile>()
-    OpenClassFilesTask({ actualClassFilesReadyToOpen.add(it) }, project)
+    OpenClassFilesTask({ actualClassFilesReadyToOpen.add(it.classFile) }, project)
       .consumeFiles(listOf(classFileAsVirtualFile))
       .openFiles()
     Assert.assertEquals(1, actualClassFilesReadyToOpen.size)
@@ -59,7 +59,7 @@ class OpenClassFilesTaskTest(
     }
 
     val actualClassFilesReadyToOpen = mutableListOf<VirtualFile>()
-    OpenClassFilesTask({ actualClassFilesReadyToOpen.add(it) }, project)
+    OpenClassFilesTask({ actualClassFilesReadyToOpen.add(it.classFile) }, project)
       .consumePsiFiles(listOf(psiJavaFile!!))
       .openFiles()
     Assert.assertEquals(1, actualClassFilesReadyToOpen.size)
@@ -130,10 +130,10 @@ class OpenClassFilesTaskTest(
 
     var filesOpened = 0
     OpenClassFilesTask({
-                         filesOpened++;
+                         filesOpened++
                          assertThat(expectedClassFileName)
-                           .describedAs("File ${it.path} should have expected class file name")
-                           .isEqualTo(it.name)
+                           .describedAs("File ${it.classFile.path} should have expected class file name")
+                           .isEqualTo(it.classFile.name)
                        }, project)
       .consumePsiElements(psiElements)
       .openFiles()